Added status codes to delete method

+++ SchoolManager.Application/Classes/Commands/DeleteClass/DeleteClassCommand.cs
@@ -1,4 +1,5 @@
 ﻿using MediatR;
+using SchoolManager.Application.Common.Exceptions;
 using SchoolManager.Application.Common.Interfaces;
 using System;
 using System.Collections.Generic;
@@ -13,7 +14,7 @@ namespace SchoolManager.Application.Classes.Commands.DeleteClass
     {
         public int ClassId { get; set; }
     }
-    public class DeleteClassCommandHandler : IRequestHandler<DeleteClassCommand>
+    public class DeleteClassCommandHandler : IRequestHandler<DeleteClassCommand,Unit>
     {
         private readonly ISchoolDbContext _context;
 
@@ -24,9 +25,16 @@ namespace SchoolManager.Application.Classes.Commands.DeleteClass
         public async Task<Unit> Handle(DeleteClassCommand request, CancellationToken cancellationToken)
         {
             var classEntity = await _context.Classes.FindAsync(request.ClassId);
+
+            if (classEntity == null) throw new NotFoundException($"Class with id: {request.ClassId} was not found");
+
             _context.Classes.Remove(classEntity);
             await _context.SaveChangesAsync();
-            return new Unit();
+
+
+            return Unit.Value;
+
+
         }
     }
 }
